+++ documents/classes/Assertable.md
@@ -1,4 +1,4 @@
-[@shellygo/cypress-test-utils - v2.0.41](../README.md) / [Modules](../modules.md) / Assertable
+[@shellygo/cypress-test-utils - v2.0.42](../README.md) / [Modules](../modules.md) / Assertable
 
 # Class: Assertable<T\>
