+++ documents/classes/CypressReactComponentHelper.md
@@ -1,4 +1,4 @@
-[@shellygo/cypress-test-utils - v2.0.41](../README.md) / [Modules](../modules.md) / CypressReactComponentHelper
+[@shellygo/cypress-test-utils - v2.0.42](../README.md) / [Modules](../modules.md) / CypressReactComponentHelper
 
 # Class: CypressReactComponentHelper
