+++ src/cypress-helper.spec.cy.ts
@@ -1,4 +1,4 @@
-import { CypressHelper, loggable } from ".";
+import { CypressHelper, loggable, match } from ".";
 import { then } from "./assertable";
 
 describe("cypress helper tests", () => {
@@ -14,6 +14,59 @@ describe("cypress helper tests", () => {
     when.visit(
       "https://htmlpreview.github.io/?https://raw.githubusercontent.com/ShellyDCMS/cypress-test-utils/main/index.html"
     );
+  });
+
+  describe("stubbing an spying", () => {
+    it("should spy on function", () => {
+      const obj = {
+        func: (param: number) => {}
+      };
+      given.spyOnObject(obj, "func");
+      obj.func(3);
+      then(get.spyFromFunction(obj.func)).shouldHaveBeenCalledWith(3);
+    });
+
+    it("should partially match spy params", () => {
+      const obj = {
+        func: (param: Object) => {}
+      };
+      given.spyOnObject(obj, "func");
+      obj.func({ shelly: "go", inner: { attr: "value" } });
+      then(get.spyFromFunction(obj.func)).shouldHaveBeenCalledWithMatch(
+        match({ inner: { attr: "value" } })
+      );
+    });
+
+    it("should stub function", () => {
+      let func = () => 5;
+      func = given.stub().returns(7);
+      expect(func()).to.eq(7);
+    });
+
+    it("should stub function with alias", () => {
+      let func = (input: number) => input;
+      func = given.stub("func");
+      func(7);
+      then(get.stub("func")).shouldHaveBeenCalledWith(7);
+    });
+
+    it("should stub object function", () => {
+      const obj = {
+        func: () => 5
+      };
+      given.stubObjectMethod(obj, "func").returns(7);
+      then(get.assertableStub(obj.func())).shouldEqual(7);
+    });
+
+    it("should stub object getter", () => {
+      const obj = {
+        get count() {
+          return 5;
+        }
+      };
+      given.stubObjectMethod(obj, "count").get(() => 7);
+      expect(obj.count).to.eq(7);
+    });
   });
 
   describe("default data attribute", () => {
