+++ src/stub-builder.spec.cy.ts
@@ -1,4 +1,4 @@
-import { CypressHelper, match } from ".";
+import { CypressHelper } from ".";
 import { then } from "./assertable";
 
 describe("stub builder tests", () => {
@@ -6,57 +6,6 @@ describe("stub builder tests", () => {
   beforeAndAfter();
   beforeEach(() => {
     ({ given, when, get } = new CypressHelper());
-  });
-
-  it("should spy on function", () => {
-    const obj = {
-      func: (param: number) => {}
-    };
-    given.spyOnObject(obj, "func");
-    obj.func(3);
-    then(get.spyFromFunction(obj.func)).shouldHaveBeenCalledWith(3);
-  });
-
-  it("should partially match spy params", () => {
-    const obj = {
-      func: (param: Object) => {}
-    };
-    given.spyOnObject(obj, "func");
-    obj.func({ shelly: "go", inner: { attr: "value" } });
-    then(get.spyFromFunction(obj.func)).shouldHaveBeenCalledWithMatch(
-      match({ inner: { attr: "value" } })
-    );
-  });
-
-  it("should stub function", () => {
-    let func = () => 5;
-    func = given.stub().returns(7);
-    expect(func()).to.eq(7);
-  });
-
-  it("should stub function with alias", () => {
-    let func = (input: number) => input;
-    func = given.stub("func");
-    func(7);
-    then(get.stub("func")).shouldHaveBeenCalledWith(7);
-  });
-
-  it("should stub object function", () => {
-    const obj = {
-      func: () => 5
-    };
-    given.stubObjectMethod(obj, "func").returns(7);
-    then(get.assertableStub(obj.func())).shouldEqual(7);
-  });
-
-  it("should stub object getter", () => {
-    const obj = {
-      get count() {
-        return 5;
-      }
-    };
-    given.stubObjectMethod(obj, "count").get(() => 7);
-    expect(obj.count).to.eq(7);
   });
 
   describe("stubbing interface", () => {
@@ -71,7 +20,8 @@ describe("stub builder tests", () => {
     }
 
     it("should stub interface", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>();
+      const mockMyInterface =
+        given.stubbedInterface<MyInterface>("MyInterface");
       mockMyInterface.func(5, "whatever");
       then(get.assertableStub(mockMyInterface.func)).shouldHaveBeenCalledWith(
         5,
@@ -80,7 +30,8 @@ describe("stub builder tests", () => {
     });
 
     it("should assert stub async function calls", async () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>();
+      const mockMyInterface =
+        given.stubbedInterface<MyInterface>("MyInterface");
       mockMyInterface.asynFunc.returns(Promise.resolve(7));
       mockMyInterface.propertyFunc(3);
       then(
@@ -90,9 +41,12 @@ describe("stub builder tests", () => {
 
     it("should assert stub async function calls using overrides", async () => {
       const stub = given.stub().returns(Promise.resolve(7));
-      const mockMyInterface = given.stubbedInterface<MyInterface>({
-        asynFunc: stub
-      });
+      const mockMyInterface = given.stubbedInterface<MyInterface>(
+        "MyInterface",
+        {
+          asynFunc: stub
+        }
+      );
       mockMyInterface.propertyFunc(3);
       then(
         get.assertableStub(mockMyInterface.asynFunc)
@@ -100,7 +54,8 @@ describe("stub builder tests", () => {
     });
 
     it("should stub async function", async () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>();
+      const mockMyInterface =
+        given.stubbedInterface<MyInterface>("MyInterface");
       mockMyInterface.asynFunc.returns(Promise.resolve(7));
       mockMyInterface.propertyFunc(3);
       expect(await mockMyInterface.asynFunc()).to.eq(7);
@@ -108,15 +63,19 @@ describe("stub builder tests", () => {
 
     it("should stub async function using overrides", async () => {
       const stub = given.stub().returns(Promise.resolve(7));
-      const mockMyInterface = given.stubbedInterface<MyInterface>({
-        asynFunc: stub
-      });
+      const mockMyInterface = given.stubbedInterface<MyInterface>(
+        "MyInterface",
+        {
+          asynFunc: stub
+        }
+      );
       mockMyInterface.propertyFunc(3);
       expect(await mockMyInterface.asynFunc()).to.eq(7);
     });
 
     it("should stub async property function", async () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>();
+      const mockMyInterface =
+        given.stubbedInterface<MyInterface>("MyInterface");
       await mockMyInterface.asyncPropertyFunc(3);
       then(
         get.assertableStub(mockMyInterface.asyncPropertyFunc)
@@ -124,13 +83,15 @@ describe("stub builder tests", () => {
     });
 
     it("should override class property Function", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>();
+      const mockMyInterface =
+        given.stubbedInterface<MyInterface>("MyInterface");
       mockMyInterface.propertyFunc.returns(7);
       expect(mockMyInterface.propertyFunc(3)).to.eq(7);
     });
 
     it("should stub interface property function", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>();
+      const mockMyInterface =
+        given.stubbedInterface<MyInterface>("MyInterface");
       mockMyInterface.propertyFunc(3);
       then(
         get.assertableStub(mockMyInterface.propertyFunc)
@@ -138,28 +99,38 @@ describe("stub builder tests", () => {
     });
 
     it("should override interface property", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>({
-        property: 5
-      });
+      const mockMyInterface = given.stubbedInterface<MyInterface>(
+        "MyInterface",
+        {
+          property: 5
+        }
+      );
       expect(mockMyInterface.property).to.eq(5);
     });
 
     it("should override interface getter", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>({
-        getter: 5
-      });
+      const mockMyInterface = given.stubbedInterface<MyInterface>(
+        "MyInterface",
+        {
+          getter: 5
+        }
+      );
       expect(mockMyInterface.getter).to.eq(5);
     });
 
     it("should stub interface setter with override", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>({
-        property: 8
-      });
+      const mockMyInterface = given.stubbedInterface<MyInterface>(
+        "MyInterface",
+        {
+          property: 8
+        }
+      );
       expect(mockMyInterface.property).to.eq(8);
     });
 
     it("should assert stub interface setter calls", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>();
+      const mockMyInterface =
+        given.stubbedInterface<MyInterface>("MyInterface");
       mockMyInterface.setter = 5;
       then(get.assertableStub(mockMyInterface.setter)).shouldHaveBeenCalledWith(
         5
@@ -167,15 +138,19 @@ describe("stub builder tests", () => {
     });
 
     it("should allow setting properties", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>({
-        property: 8
-      });
+      const mockMyInterface = given.stubbedInterface<MyInterface>(
+        "MyInterface",
+        {
+          property: 8
+        }
+      );
       mockMyInterface.property = 5;
       expect(mockMyInterface.property).to.eq(8);
     });
 
     it("should stub interface function return value", () => {
-      const mockMyInterface = given.stubbedInterface<MyInterface>();
+      const mockMyInterface =
+        given.stubbedInterface<MyInterface>("MyInterface");
       mockMyInterface.func.returns(7);
       expect(mockMyInterface.func(5, "whatever")).to.eq(7);
     });
